post

+++ app/Http/Controllers/TasksController.php
@@ -17,11 +17,20 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('id')->paginate(1000);
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            // $data += $this->counts($user);
+            return view('tasks.index', $data);
+        }else {
+            return view('welcome');
+        }
     }
 
     /**
@@ -47,17 +56,21 @@ class TasksController extends Controller
     public function store(Request $request)
     {
         
-        $this->validate($request, [
-            'status' => 'required|max:10',
+       $this->validate($request, [
             'content' => 'required|max:191',
+            'status' => 'required|max:10',
+        ]);
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
         ]);
         
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-
-        return redirect('/');
+        // $task = new Task;
+        // $task->status = $request->status; 
+        // $task->content = $request->content;
+        // $task->save();
+        
+        return redirect('/tasks');
     }
 
     /**
@@ -69,10 +82,14 @@ class TasksController extends Controller
     public function show($id)
     {
         $task = Task::find($id);
-
-        return view('tasks.show', [
+        
+        if (\Auth::user()->id === $task->user_id){
+            return view('tasks.show', [
             'task' => $task,
         ]);
+        } else {
+        return redirect('/');
+        }
     }
 
     /**
@@ -84,10 +101,14 @@ class TasksController extends Controller
     public function edit($id)
     {
         $task = Task::find($id);
-
-        return view('tasks.edit', [
+        
+        if (\Auth::user()->id === $task->user_id){
+            return view('tasks.edit', [
             'task' => $task,
         ]);
+        return redirect('/');
+            
+        }
     }
 
     /**
@@ -121,9 +142,13 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
-
+        $task = \App\Task::find($id);
+        if (\Auth::user()->id === $task->user_id) {
+            $task->delete();
         return redirect('/');
+        } else {
+        //return redirect()->back();
+        return view('welcome');
+        }
     }
 }
